+++ aggreg_operators/minaggregoperator.cpp
@@ -4,7 +4,7 @@
 
 #include <QDebug>
 
-int MinAggregOperator::GetWorthyValue(const QList<long double> &t_list)
+int MinAggregOperator::GetWorthyValue(const QList<double>& t_list)
 {
     if ( true == t_list.isEmpty() )
     {
